@@ -2,69 +2,127 @@
 
 namespace emailboilerplateforatkdata;
 
-use Atk4\Data\Exception;
+use Atk4\Data\Persistence;
 use Atk4\Ui\HtmlTemplate;
+use DirectoryIterator;
+use Throwable;
 
 class EmailTemplateLoader
 {
 
     public static function getEmailTemplate(BaseEmail $baseEmail): HtmlTemplate
     {
-
+        //try load Template for the individual $baseEmail->model entity
+        $result = self::tryLoadTemplateForEntity($baseEmail);
+        if ($result) {
+            return $result;
+        }
+        //try load existing Template from persistence
+        $result = self::tryLoadTemplateFromPersistence($baseEmail);
+        if ($result) {
+            return $result;
+        }
+        //load default template from file if other methods did not return anything
+        return self::loadDefaultTemplateFromFile($baseEmail);
     }
 
-    public function loadEmailTemplate(string $name, bool $raw_template = false, array $customFromModels = [])
+    protected static function tryLoadTemplateForEntity(BaseEmail $baseEmail): ?HtmlTemplate
     {
-        $template = new Template();
-        $template->app = $this;
-        //try to load From EmailTemplate per Model
-        $et = $this->_getCustomEmailTemplateFromModel($name, $customFromModels);
-        //else try to load from DB
-        if (!$et) {
-            $et = new EmailTemplate($this->db);
-            $et->addCondition('model_class', '=', null);
-            $et->addCondition('model_id', '=', null);
-            $et->tryLoadBy('ident', $name);
+        //no model or not loaded?
+        if (!$baseEmail->model || !$baseEmail->model->loaded()) {
+            return null;
         }
 
-        if ($et->loaded()) {
-            if ($raw_template) {
-                return $et->get('value');
-            } else {
-                $template->loadTemplateFromString((string)$et->get('value'));
-                return $template;
-            }
-        }
-
-        //now try to load from file
-        $fileName = FILE_BASE_PATH . $this->emailTemplateDir . '/' . $name;
-        if (file_exists($fileName)) {
-            if ($raw_template) {
-                return file_get_contents($fileName);
-            } elseif ($t = $template->tryLoad($fileName)) {
-                return $t;
-            }
-        }
-
-        throw new Exception('Can not find email template file: ' . $name);
+        return self::customLoadTemplateForEntity($baseEmail);
     }
 
-    protected function _getCustomEmailTemplateFromModel(string $name, array $customFromModels): ?EmailTemplate
+    //overwrite for custom implementations
+    protected static function customLoadTemplateForEntity(BaseEmail $baseEmail): ?HtmlTemplate
     {
-        foreach ($customFromModels as $model) {
-            if (!$model->loaded()) {
-                throw new Exception('Model needs to be loaded in ' . __FUNCTION__);
-            }
-            $et = new EmailTemplate($this->db);
-            $et->addCondition('model_class', get_class($model));
-            $et->addCondition('model_id', $model->get('id'));
-            $et->tryLoadBy('ident', $name);
-            if ($et->loaded()) {
-                return clone $et;
-            }
-        }
-
         return null;
+    }
+
+    protected static function tryLoadTemplateFromPersistence(BaseEmail $baseEmail): ?HtmlTemplate
+    {
+        $emailTemplate = new EmailTemplate($baseEmail->persistence);
+        $emailTemplate->addCondition('model_class', '=', null);
+        $emailTemplate->addCondition('model_id', '=', null);
+        $emailTemplate->tryLoadBy('ident', (new \ReflectionClass($baseEmail))->getName());
+        if (!$emailTemplate->loaded()) {
+            return null;
+        }
+        $htmlTemplate = new HtmlTemplate($emailTemplate->get('value'));
+        return $htmlTemplate;
+    }
+
+    protected static function loadDefaultTemplateFromFile(BaseEmail $baseEmail): HtmlTemplate
+    {
+        //now try to load from file
+        $fileName = self::getTemplateFilePath($baseEmail);
+        //throws Exception if file can not be found
+        $htmlTemplate = new HtmlTemplate();
+        $htmlTemplate->loadFromFile($fileName);
+        return $htmlTemplate;
+    }
+    protected static function loadRawDefaultTemplateFromFile(BaseEmail $baseEmail): string
+    {
+        //now try to load from file
+        $fileName = self::getTemplateFilePath($baseEmail);
+        return file_get_contents($fileName);
+    }
+
+    //overwrite in custom implementations to easily define where default templates can be found
+    protected static function getTemplateFilePath(BaseEmail $baseEmail): string
+    {
+        return $baseEmail->defaultTemplateFile;
+    }
+
+    public static function createEmailTemplateEntities(array $dirs, Persistence $persistence): void
+    {
+        foreach (self:: getAllBaseEmailImplementations($dirs, $persistence) as $className => $baseEmail) {
+            $emailTemplate = new EmailTemplate($persistence);
+            $emailTemplate->addCondition('model_class', null);
+            $emailTemplate->addCondition('model_id', null);
+            $emailTemplate->tryLoadBy('ident', (new \ReflectionClass($baseEmail))->getName());
+            if (!$emailTemplate->loaded()) {
+                $emailTemplate->set('ident', $baseEmail->template);
+                $emailTemplate->set('value', self::loadRawDefaultTemplateFromFile($baseEmail));
+                $emailTemplate->save();
+            }
+        }
+    }
+
+    /**
+     * return an instance of each found implementation of BaseEmail in the given folder(s)
+     * parameter array: key is the dir to check for classes, value is the namespace
+     */
+    protected static function getAllBaseEmailImplementations(array $dirs, Persistence $persistence): array
+    {
+        $result = [];
+
+        foreach ($dirs as $dir => $namespace) {
+            foreach (new DirectoryIterator($dir) as $file) {
+                if ($file->getExtension() !== 'php') {
+                    continue;
+                }
+                $className = $namespace . $file->getBasename('.php');
+                if (!class_exists($className)) {
+                    continue;
+                }
+                try {
+                    $instance = new $className($persistence);
+                } catch (Throwable $e) {
+                    continue;
+                }
+                if (!$instance instanceof BaseEmail) {
+                    continue;
+                }
+
+                $result[$className] = clone $instance;
+            }
+        }
+
+        return $result;
     }
 
     public function saveEmailTemplate(
@@ -91,6 +149,4 @@ class EmailTemplateLoader
 
         return $emailTemplate;
     }
-
-
 }
